# packages/cactus-plugin-carbon-credit/src/main/go/generated/openapi/go-client/api/openapi.yaml
openapi: 3.0.3
info:
  description: Implementation for Carbon Credit Plugin logic
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  title: Hyperledger Cactus Plugin - Carbon Credit
  version: 1.0.0
servers:
- url: /
paths:
  /api/v1/@hyperledger/cactus-plugin-carbon-credit/specific-buy:
    post:
      description: Acquire a specific basket of TCO2s from a marketplace.
      operationId: specificBuyRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecificBuyRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpecificBuyResponse'
          description: Successful response
      x-hyperledger-cacti:
        http:
          verbLowerCase: post
          path: /api/v1/@hyperledger/cactus-plugin-carbon-credit/specific-buy
  /api/v1/@hyperledger/cactus-plugin-carbon-credit/random-buy:
    post:
      description: Acquire a random basket of TCO2s from a marketplace.
      operationId: randomBuyRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RandomBuyRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RandomBuyResponse'
          description: Successful response
      x-hyperledger-cacti:
        http:
          verbLowerCase: post
          path: /api/v1/@hyperledger/cactus-plugin-carbon-credit/random-buy
  /api/v1/@hyperledger/cactus-plugin-carbon-credit/retire:
    post:
      description: Retire a specific basket of TCO2s on a marketplace.
      operationId: retireRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RetireRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetireResponse'
          description: Successful response
      x-hyperledger-cacti:
        http:
          verbLowerCase: post
          path: /api/v1/@hyperledger/cactus-plugin-carbon-credit/retire
  /api/v1/@hyperledger/cactus-plugin-carbon-credit/get-available-tco2s:
    post:
      description: Get a list of available TCO2s (Tokenized Carbon Offset Units).
      operationId: getAvailableTCO2sRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetAvailableTCO2sRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetAvailableTCO2sResponse'
          description: Successful response with a list of available VCUs.
      x-hyperledger-cacti:
        http:
          verbLowerCase: post
          path: /api/v1/@hyperledger/cactus-plugin-carbon-credit/get-available-tco2s
  /api/v1/@hyperledger/cactus-plugin-carbon-credit/get-vcu-metadata:
    post:
      description: Get metadata for a specific VCU.
      operationId: getVCUMetadataRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetVCUMetadataRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VCUMetadata'
          description: Successful response with VCU metadata.
      x-hyperledger-cacti:
        http:
          verbLowerCase: post
          path: /api/v1/@hyperledger/cactus-plugin-carbon-credit/get-vcu-metadata
  /api/v1/@hyperledger/cactus-plugin-carbon-credit/get-purchase-price:
    post:
      description: Get the purchase price for a specific VCU.
      operationId: getPurchasePriceRequest
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GetPurchasePriceRequest'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetPurchasePriceResponse'
          description: Successful response with VCU metadata.
      x-hyperledger-cacti:
        http:
          verbLowerCase: post
          path: /api/v1/@hyperledger/cactus-plugin-carbon-credit/get-purchase-price
components:
  schemas:
    Marketplace:
      enum:
      - Toucan
      type: string
    Network:
      enum:
      - polygon
      - ethereum
      - celo
      - alfajores
      - base
      - base-sepolia
      type: string
    NetworkConfig:
      properties:
        rpcUrl:
          type: string
        network:
          $ref: '#/components/schemas/Network'
      required:
      - network
      - rpcUrl
      type: object
    SpecificBuyRequest:
      example:
        paymentToken: paymentToken
        marketplace: null
        items:
          key: 0.8008281904610115
        walletObject: walletObject
        network: null
      properties:
        marketplace:
          $ref: '#/components/schemas/Marketplace'
        network:
          $ref: '#/components/schemas/Network'
        paymentToken:
          type: string
        items:
          additionalProperties:
            type: number
          description: A mapping from address (string) to amount (number).
          type: object
        walletObject:
          type: string
      required:
      - items
      - marketplace
      - network
      - paymentToken
      - walletObject
      type: object
    SpecificBuyResponse:
      example:
        assetAmount: assetAmount
        txHashSwap: txHashSwap
        tco2List:
        - amount: amount
          address: address
        - amount: amount
          address: address
      properties:
        txHashSwap:
          type: string
        assetAmount:
          type: string
        tco2List:
          items:
            $ref: '#/components/schemas/SpecificBuyResponse_tco2List_inner'
          type: array
      required:
      - assetAmount
      - txHashSwap
      type: object
    RandomBuyRequest:
      example:
        paymentToken: paymentToken
        amount: 0.8008281904610115
        marketplace: null
        walletObject: walletObject
        network: null
      properties:
        marketplace:
          $ref: '#/components/schemas/Marketplace'
        network:
          $ref: '#/components/schemas/Network'
        paymentToken:
          type: string
        amount:
          type: number
        walletObject:
          type: string
      required:
      - amount
      - marketplace
      - network
      - walletObject
      type: object
    RandomBuyResponse:
      example:
        assetAmount: assetAmount
        txHashSwap: txHashSwap
        tco2List:
        - amount: amount
          address: address
        - amount: amount
          address: address
      properties:
        txHashSwap:
          type: string
        assetAmount:
          type: string
        tco2List:
          items:
            $ref: '#/components/schemas/SpecificBuyResponse_tco2List_inner'
          type: array
      required:
      - assetAmount
      - txHashSwap
      type: object
    RetireRequest:
      example:
        tco2s:
        - tco2s
        - tco2s
        retirementReason: retirementReason
        marketplace: null
        amounts:
        - 0.8008281904610115
        - 0.8008281904610115
        beneficiaryName: beneficiaryName
        beneficiaryAddress: beneficiaryAddress
        message: message
        walletObject: walletObject
        network: null
      properties:
        marketplace:
          $ref: '#/components/schemas/Marketplace'
        network:
          $ref: '#/components/schemas/Network'
        walletObject:
          type: string
        tco2s:
          items:
            type: string
          type: array
        amounts:
          items:
            type: number
          type: array
        beneficiaryAddress:
          type: string
        beneficiaryName:
          type: string
        message:
          type: string
        retirementReason:
          type: string
      required:
      - amounts
      - beneficiaryAddress
      - beneficiaryName
      - marketplace
      - message
      - network
      - retirementReason
      - tco2s
      - walletObject
      type: object
    RetireResponse:
      example:
        txHashRetire: txHashRetire
        retirementCertificate: retirementCertificate
      properties:
        txHashRetire:
          type: string
        retirementCertificate:
          type: string
      required:
      - retirementCertificate
      - txHashRetire
      type: object
    GetAvailableTCO2sRequest:
      example:
        filterCriteria: filterCriteria
        marketplace: null
        limit: limit
        orderBy: orderBy
        network: null
      properties:
        marketplace:
          $ref: '#/components/schemas/Marketplace'
        network:
          $ref: '#/components/schemas/Network'
        filterCriteria:
          type: string
        orderBy:
          type: string
        limit:
          type: string
      required:
      - marketplace
      - network
      type: object
    GetAvailableTCO2sResponse:
      example:
        totalCount: 0.8008281904610115
        tco2List:
        - tco2List
        - tco2List
      properties:
        tco2List:
          items:
            type: string
          type: array
        totalCount:
          type: number
      required:
      - tco2List
      - totalCount
      type: object
    GetVCUMetadataRequest:
      example:
        marketplace: null
        vcuIdentifier: vcuIdentifier
        projectIdentifier: projectIdentifier
        network: null
      properties:
        marketplace:
          $ref: '#/components/schemas/Marketplace'
        network:
          $ref: '#/components/schemas/Network'
        projectIdentifier:
          type: string
        vcuIdentifier:
          type: string
      required:
      - marketplace
      - network
      - projectIdentifier
      - vcuIdentifier
      type: object
    VCUMetadata:
      example:
        symbol: symbol
        totalSupply: 0.8008281904610115
        name: name
        attributes:
          other: other
          vintage: vintage
          region: region
          methodology: methodology
      properties:
        name:
          type: string
        symbol:
          type: string
        totalSupply:
          type: number
        attributes:
          $ref: '#/components/schemas/VCUMetadata_attributes'
      required:
      - attributes
      - name
      - symbol
      - totalSupply
      type: object
    GetPurchasePriceRequest:
      example:
        unit: unit
        amount: 0.8008281904610115
        platform: null
        network: null
      properties:
        platform:
          $ref: '#/components/schemas/Marketplace'
        network:
          $ref: '#/components/schemas/Network'
        unit:
          description: The token address to retrieve the price for.
          type: string
        amount:
          description: The amount of Units to price.
          type: number
      required:
      - amount
      - network
      - platform
      - unit
      type: object
    GetPurchasePriceResponse:
      example:
        price: price
      properties:
        price:
          description: The current price in USDC for the requested amount.
          type: string
      required:
      - price
      type: object
    SpecificBuyResponse_tco2List_inner:
      example:
        amount: amount
        address: address
      properties:
        address:
          type: string
        amount:
          type: string
      required:
      - address
      - amount
      type: object
    VCUMetadata_attributes:
      example:
        other: other
        vintage: vintage
        region: region
        methodology: methodology
      properties:
        region:
          type: string
        vintage:
          type: string
        methodology:
          type: string
        other:
          type: string
      type: object