# packages/cactus-plugin-carbon-credit/src/main/typescript/web-services/retire-endpoint.ts
import { Express, Request, Response } from "express";

import {
  Logger,
  Checks,
  LogLevelDesc,
  LoggerProvider,
  IAsyncProvider,
  safeStringifyException,
} from "@hyperledger/cactus-common";
import {
  IEndpointAuthzOptions,
  IExpressRequestHandler,
  IWebServiceEndpoint,
} from "@hyperledger/cactus-core-api";
import { registerWebServiceEndpoint } from "@hyperledger/cactus-core";

import {
  RetireRequest,
  RetireResponse,
} from "./../generated/openapi/typescript-axios";

import { PluginCarbonCredit } from "../plugin-carbon-credit";

import OAS from "../../json/openapi.json";

export interface IRetireEndpointOptions {
  logLevel?: LogLevelDesc;
  plugin: PluginCarbonCredit;
}

export class RetireEndpoint implements IWebServiceEndpoint {
  public static readonly CLASS_NAME = "RetireEndpoint";

  private readonly log: Logger;

  public get className(): string {
    return RetireEndpoint.CLASS_NAME;
  }

  constructor(public readonly options: IRetireEndpointOptions) {
    const fnTag = `${this.className}#constructor()`;
    Checks.truthy(options, `${fnTag} arg options`);
    Checks.truthy(options.plugin, `${fnTag} arg options.plugin`);

    const level = options.logLevel || "INFO";
    const label = this.className;
    this.log = LoggerProvider.getOrCreate({ level, label });
  }

  public get oasPath(): (typeof OAS.paths)["/api/v1/@hyperledger/cactus-plugin-carbon-credit/retire"] {
    return OAS.paths["/api/v1/@hyperledger/cactus-plugin-carbon-credit/retire"];
  }

  public getPath(): string {
    return this.oasPath.post["x-hyperledger-cacti"].http.path;
  }

  public getVerbLowerCase(): string {
    return this.oasPath.post["x-hyperledger-cacti"].http.verbLowerCase;
  }

  public getOperationId(): string {
    return this.oasPath.post.operationId;
  }

  getAuthorizationOptionsProvider(): IAsyncProvider<IEndpointAuthzOptions> {
    // TODO: make this an injectable dependency in the constructor
    return {
      get: async () => ({
        isProtected: true,
        requiredRoles: [],
      }),
    };
  }

  public async registerExpress(
    expressApp: Express,
  ): Promise<IWebServiceEndpoint> {
    await registerWebServiceEndpoint(expressApp, this);
    return this;
  }

  public getExpressRequestHandler(): IExpressRequestHandler {
    return this.handleRequest.bind(this);
  }

  private async handleRequest(req: Request, res: Response): Promise<void> {
    const reqTag = `${this.getVerbLowerCase()} - ${this.getPath()}`;
    this.log.debug(reqTag);

    const reqBody: RetireRequest = req.body;
    const {
      marketplace,
      walletObject,
      tco2s,
      amounts,
      beneficiaryAddress,
      beneficiaryName,
      message,
      retirementReason,
    } = reqBody;

    if (
      !marketplace ||
      !walletObject ||
      !tco2s ||
      !amounts ||
      !beneficiaryAddress ||
      !beneficiaryName ||
      !message ||
      !retirementReason
    ) {
      const errorMessage = "Missing required parameters for retirement.";
      this.log.error(errorMessage);
      res.status(400).json({ error: errorMessage });
      return;
    }

    try {
      this.log.info(
        `Received a retire request for ${amounts} units on marketplace ${marketplace}`,
      );
      const retireResponse: RetireResponse =
        await this.options.plugin.retire(reqBody);
      res.status(200).json(retireResponse);
    } catch (ex) {
      this.log.error(`Crash while serving ${reqTag}`, ex);
      res.status(500).json({
        message: "Internal Server Error",
        error: safeStringifyException(ex),
      });
    }
  }
}